requirements.txt
streamlit>=1.28.0
pandas>=2.2.0
plotly>=5.17.0
requests>=2.31.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.0
google-api-python-client>=2.108.0
gspread>=5.12.0
python-dotenv>=1.0.0
altair>=5.1.0
numpy>=1.24.0
openpyxl>=3.1.0
xlsxwriter>=3.1.0
